+++ database/migrations/2025_05_29_013206_add_is_void_to_payments_table.php
@@ -19,7 +19,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('payments', function (Blueprint $table) {
-            $table->dropColumn('is_void');
+            if (Schema::hasColumn('payments', 'is_void')) {
+                $table->dropColumn('is_void');
+            }
         });
     }
 };
